+++ g2g-plugins/src/lib.rs
@@ -9,6 +9,7 @@
 
 extern crate alloc;
 
+pub mod audioconvert;
 pub mod audiotestsrc;
 pub mod capsfilter;
 pub mod fakesink;